requirements.txt
appdirs==1.4.4
asgiref==3.8.1
attrs==23.2.0
Babel==2.15.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
Bottleneck==1.3.8
Brotli==1.1.0
certifi==2024.7.4
chardet==5.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.8
cssselect==1.2.0
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
Django==5.0.7
django-cors-headers==4.3.1
djangorestframework==3.15.2
et-xmlfile==1.1.0
fonttools==4.53.1
html5lib==1.1
idna==3.7
Jinja2==3.1.4
jsonschema==4.22.0
kiwisolver==1.4.5
lxml==5.2.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mutagen==1.47.0
nbformat==5.10.4
netaddr==1.3.0
numexpr==2.10.1
numpy==2.0.1
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
paramiko==3.4.0
pillow==10.4.0
platformdirs==4.2.2
py-cpuinfo==9.0.0
pycryptodomex==3.20.0
Pygments==2.18.0
PyJWT==2.9.0
PyNaCl==1.5.0
PyOpenGL==3.1.7
pyparsing==3.1.2
pyqtgraph==0.13.7
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rich==13.7.1
scipy==1.14.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
sympy==1.13.1
tables==3.9.2
traitlets==5.14.3
typing_extensions==4.12.2
ufoLib2==0.16.0
unicodedata2==15.1.0
urllib3==2.2.2
webencodings==0.5.1
websockets==12.0
wheel==0.43.0
yt-dlp==2024.8.6
